refactor: Converter를 infrastructure 캐시지 내부로 이동
